refactor: extract suggestedFilters
