+++ src/schema/v2/previewSavedSearch/previewSavedSearch.ts
@@ -8,10 +8,6 @@ import {
   GraphQLObjectType,
   GraphQLString,
 } from "graphql"
-import artworkMediums from "lib/artworkMediums"
-import attributionClasses from "lib/attributionClasses"
-import { isExisty } from "lib/helpers"
-import _ from "lodash"
 import { ResolverContext } from "types/graphql"
 import ArtworkSizes from "../artwork/artworkSizes"
 import {
@@ -20,6 +16,7 @@ import {
 } from "./searchCriteriaLabel"
 import { generateDisplayName } from "./generateDisplayName"
 import { resolveHref } from "./resolveHref"
+import { suggestedFilters } from "./suggestedFilters"
 
 const previewSavedSearchArgs: GraphQLFieldConfigArgumentMap = {
   acquireable: {
@@ -76,8 +73,6 @@ const previewSavedSearchArgs: GraphQLFieldConfigArgumentMap = {
   },
 }
 
-const ALLOWED_RARITY_SUGGESTIONS = ["unique", "limited edition"]
-
 const PreviewSavedSearchType = new GraphQLObjectType<any, ResolverContext>({
   name: "PreviewSavedSearch",
   fields: () => ({
@@ -105,98 +100,10 @@ const PreviewSavedSearchType = new GraphQLObjectType<any, ResolverContext>({
       type: new GraphQLNonNull(
         new GraphQLList(new GraphQLNonNull(SearchCriteriaLabel))
       ),
-      resolve: async ({ artistIDs }, _args, { filterArtworksLoader }) => {
-        if (!artistIDs) {
-          throw new Error("artistIDs are required to get suggested filters")
-        }
-
-        const gravityArgs = {
-          size: 0,
-          aggregations: ["attribution_class", "medium"],
-          artist_ids: artistIDs,
-        }
-
-        const { aggregations } = await filterArtworksLoader(gravityArgs)
-
-        const suggestedFilters: SearchCriteriaLabel[] = []
-
-        const rarityOptions = _.chain(
-          getMostPopularOptions(aggregations["attribution_class"])
-        )
-          .map((label) => getRaritySearchCriteriaLabel(label))
-          .compact()
-          .filter((rarity) => ALLOWED_RARITY_SUGGESTIONS.includes(rarity.value))
-          .value()
-
-        if (rarityOptions.length) {
-          suggestedFilters.push(...rarityOptions)
-        }
-
-        const mediumOptions = _.chain(
-          getMostPopularOptions(aggregations["medium"])
-        )
-          .map((label) => getMediumSearchCriteriaLabel(label))
-          .compact()
-          .value()
-
-        if (mediumOptions.length) {
-          suggestedFilters.push(...mediumOptions)
-        }
-
-        return suggestedFilters
-      },
+      resolve: suggestedFilters,
     },
   }),
 })
-
-const getMostPopularOptions = (aggregation: {
-  [key: string]: { name: string; count: number }
-}): string[] => {
-  if (isExisty(aggregation)) {
-    const keys = Object.keys(aggregation).slice(0, 2)
-    return keys
-  }
-
-  return []
-}
-
-const getRaritySearchCriteriaLabel = (
-  value: string | undefined
-): SearchCriteriaLabel | null => {
-  if (!value) {
-    return null
-  }
-
-  return {
-    displayValue: attributionClasses[value].name,
-    field: "attributionClass",
-    value,
-    name: "Rarity",
-  }
-}
-
-const getMediumSearchCriteriaLabel = (
-  value: string | undefined
-): SearchCriteriaLabel | null => {
-  if (!value) {
-    return null
-  }
-
-  const mediumData = Object.values(artworkMediums).find((artworkMedium) => {
-    return artworkMedium.mediumFilterGeneSlug === value
-  })
-
-  if (!mediumData || !mediumData.mediumFilterGeneSlug) {
-    return null
-  }
-
-  return {
-    displayValue: mediumData.name,
-    field: "additionalGeneIDs",
-    value: mediumData.mediumFilterGeneSlug,
-    name: "Medium",
-  }
-}
 
 const PreviewSavedSearchAttributesType = new GraphQLInputObjectType({
   name: "PreviewSavedSearchAttributes",
